create functional component MuiVirtualizedTable
